nested_serializers_project/requirements.txt
arrow==0.15.6
asgiref==3.2.10
blessed==1.17.8
Django==3.0.7
django-filter==2.3.0
django-filters==0.2.1
django-picklefield==3.0.1
django-q==1.2.4
djangorestframework==3.11.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
wcwidth==0.2.4
